+++ app/Http/Controllers/FilmsController.php
@@ -22,8 +22,12 @@ class FilmsController extends Controller
      */
     public function index()
     {
-        $films = Films::all();
-        return view('film/index', ['films' => $films]);
+        if (Auth::check() && Auth::user()->is_admin == 1) {
+            $films = Films::all();
+            return view('film/index', ['films' => $films]);
+        } else {
+            return redirect(route('out'));
+        }
     }
 
     /**
@@ -33,10 +37,14 @@ class FilmsController extends Controller
      */
     public function create()
     {
-        $allActors = Actors::all();
-        $actors = new ActorsFilms();
-        $film = new Films;
-        return view('film/create', compact('film', 'allActors', 'actors'));
+        if (Auth::check() && $this->authorize('create', Films::class)) {
+            $allActors = Actors::all();
+            $actors = new ActorsFilms();
+            $film = new Films;
+            return view('film/create', compact('film', 'allActors', 'actors'));
+        } else {
+            return redirect(route('out'));
+        }
     }
 
     /**
@@ -100,6 +108,7 @@ class FilmsController extends Controller
 
 
         return redirect(route('films.index'))->with('successMsg', 'Student Successfully Added');
+
     }
 
     /**
@@ -110,27 +119,30 @@ class FilmsController extends Controller
      */
     public function edit($id)
     {
-
-        if ($film = Films::find($id)) {
-            $af = Films::find($id);//actors for this film
-            $imageArray = array();
-            $videoArray = array();
-            $actorsArray = array();
-            foreach ($af->actors as $a) {
-                array_push($actorsArray, $a->name);
-            }
-            foreach ($af->media as $b) {
-                $ext = pathinfo($b->path, PATHINFO_EXTENSION);
-                if ($ext == 'jpg' || $ext == 'jpeg' || $ext == 'png' || $ext == 'gif') {
-                    array_push($imageArray, $b->path);
-                } else {
-                    array_push($videoArray, $b->path);
+        if (Auth::check() && $this->authorize('update', Films::find($id))) {
+            if ($film = Films::find($id)) {
+                $af = Films::find($id);//actors for this film
+                $imageArray = array();
+                $videoArray = array();
+                $actorsArray = array();
+                foreach ($af->actors as $a) {
+                    array_push($actorsArray, $a->name);
                 }
+                foreach ($af->media as $b) {
+                    $ext = pathinfo($b->path, PATHINFO_EXTENSION);
+                    if ($ext == 'jpg' || $ext == 'jpeg' || $ext == 'png' || $ext == 'gif') {
+                        array_push($imageArray, $b->path);
+                    } else {
+                        array_push($videoArray, $b->path);
+                    }
+                }
+                $allActors = Actors::all();
+                return view('film/edit', compact('film', 'imageArray', 'allActors', 'actorsArray', 'videoArray'));
+            } else {
+                return redirect(route('films.index'))->with('errorMsg', 'This ID is not exist please try again');
             }
-            $allActors = Actors::all();
-            return view('film/edit', compact('film', 'imageArray', 'allActors', 'actorsArray', 'videoArray'));
         } else {
-            return redirect(route('films.index'))->with('errorMsg', 'This ID is not exist please try again');
+            return redirect(route('out'));
         }
     }
 
@@ -144,44 +156,50 @@ class FilmsController extends Controller
     public function update(Request $request, $id)
     {
         $film = Films::find($id);
-        if ($film) {
-            $film->admin_id = Auth::user()->id;
-            $film->name = $request->name;
-            $film->summary = $request->information;
-            $film->save();
-            $currentId = $film->id;
-            //get actors and store it
-            if ($request->actors) {
-                foreach ($request->actors as $actor) {
+        if (Auth::check() && $this->authorize('update', Films::find($id))) {
+
+            if ($film) {
+                $film->admin_id = Auth::user()->id;
+                $film->name = $request->name;
+                $film->summary = $request->information;
+                $film->save();
+                $currentId = $film->id;
+                //get actors and store it
+                if ($request->actors) {
+                    foreach ($request->actors as $actor) {
+
+                    }
+                }
+                //get image and store it
+                $fm = new FilmsMedia;
+                if ($request->file('images')) {
+                    foreach ($request->file('images') as $image) {
+                        $fi = Films::find($currentId);
+                        $input['imageName'] = helper::storeImage($image, '/films_thumbnail', '/films_uploads');
+                        $fm = new FilmsMedia(['admin_id' => Auth::user()->id, 'path' => $input['imageName']]);
+                        $fi->media()->save($fm);
+                    }
+                }
+                //get trailers and store it
+                if ($request->file('trailers')) {
+                    foreach ($request->file('trailers') as $trialer) {
+                        $input['trialName'] = helper::storeTrailers($trialer, '/films_trials');
+                        $fi = Films::find($currentId);
+                        $fm = new FilmsMedia(['admin_id' => Auth::user()->id, 'path' => $input['trialName']]);
+                        $fi->media()->save($fm);
+
+                    }
 
                 }
-            }
-            //get image and store it
-            $fm = new FilmsMedia;
-            if ($request->file('images')) {
-                foreach ($request->file('images') as $image) {
-                    $fi = Films::find($currentId);
-                    $input['imageName'] = helper::storeImage($image, '/films_thumbnail', '/films_uploads');
-                    $fm = new FilmsMedia(['admin_id' => Auth::user()->id, 'path' => $input['imageName']]);
-                    $fi->media()->save($fm);
-                }
-            }
-            //get trailers and store it
-            if ($request->file('trailers')) {
-                foreach ($request->file('trailers') as $trialer) {
-                    $input['trialName'] = helper::storeTrailers($trialer, '/films_trials');
-                    $fi = Films::find($currentId);
-                    $fm = new FilmsMedia(['admin_id' => Auth::user()->id, 'path' => $input['trialName']]);
-                    $fi->media()->save($fm);
+                return redirect(route('films.index'))->with('successMsg', 'Films Successfully updated');
 
-                }
-
+            } else {
+                return redirect(route('films.index'))->with('errorMsg', 'This Film iss not exist ');
             }
-            return redirect(route('films.index'))->with('successMsg', 'Films Successfully updated');
-
-        } else {
-            return redirect(route('films.index'))->with('errorMsg', 'This Film iss not exist ');
+        }else{
+            redirect(route('out'));
         }
+
     }
 
 
@@ -193,10 +211,14 @@ class FilmsController extends Controller
      */
     public function destroy($id)
     {
-        if (Films::find($id)->delete()) {
-            return redirect(route('films.index'))->with('successMsg', 'Films Successfully Delete');
+        if (Auth::check() && $this->authorize('delete', Films::find($id))) {
+            if (Films::find($id)->delete()) {
+                return redirect(route('films.index'))->with('successMsg', 'Films Successfully Delete');
+            } else {
+                return redirect(route('films.index'))->with('errorMsg', 'Something Error please try again');
+            }
         } else {
-            return redirect(route('films.index'))->with('errorMsg', 'Something Error please try again');
+            return redirect(route('out'));
         }
     }
